+++ packages/react-server/package.json
@@ -47,9 +47,10 @@
     "vitefu": "^1.0.5"
   },
   "devDependencies": {
-    "@hiogawa/vite-rsc": "workspace:*",
     "@edge-runtime/cookies": "^6.0.0",
-    "@tanstack/history": "^1.99.13"
+    "@hiogawa/vite-rsc": "workspace:*",
+    "@tanstack/history": "^1.99.13",
+    "rsc-html-stream": "^0.0.6"
   },
   "peerDependencies": {
     "react": "*",
